+++ MvvmCross.Droid.Support.V7.Fragging/Caching/MvxCachingFragmentActivity.cs
@@ -21,8 +21,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using MvvmCross.Droid.Support.V7.Fragging.Caching;
+using MvvmCross.Droid.Support.V7.Fragging.Attributes;
 
-namespace MvvmCross.Droid.Support.V7.Fragging
+namespace MvvmCross.Droid.Support.V7.Fragging.Caching
 {
     public class MvxCachingFragmentActivity
         : MvxFragmentActivity, IFragmentCacheableActivity
