+++ Samples/Example.Core/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="MvvmCross.HotTuna.Binding" version="4.0.0-beta5" targetFramework="portable-net45+win+wp80+MonoTouch10+MonoAndroid10+xamarinmac20+xamarinios10" />
-  <package id="MvvmCross.HotTuna.CrossCore" version="4.0.0-beta5" targetFramework="portable-net45+win+wp80+MonoTouch10+MonoAndroid10+xamarinmac20+xamarinios10" />
-  <package id="MvvmCross.HotTuna.MvvmCrossLibraries" version="4.0.0-beta5" targetFramework="portable-net45+win+wp80+MonoTouch10+MonoAndroid10+xamarinmac20+xamarinios10" />
+  <package id="MvvmCross.HotTuna.Binding" version="4.0.0-beta7" targetFramework="portable-net45+win+wp80+MonoTouch10+MonoAndroid10+xamarinmac20+xamarinios10" />
+  <package id="MvvmCross.HotTuna.CrossCore" version="4.0.0-beta7" targetFramework="portable-net45+win+wp80+MonoTouch10+MonoAndroid10+xamarinmac20+xamarinios10" />
+  <package id="MvvmCross.HotTuna.MvvmCrossLibraries" version="4.0.0-beta7" targetFramework="portable-net45+win+wp80+MonoTouch10+MonoAndroid10+xamarinmac20+xamarinios10" />
 </packages>
