+++ Samples/Example.Droid/Bootstrap/JsonPluginBootstrap.cs
