+++ Samples/Example.Droid/Bootstrap/VisibilityPluginBootstrap.cs
