+++ Samples/Example.Droid/Fragments/MenuFragment.cs
@@ -5,7 +5,7 @@ using Android.Runtime;
 using Android.Support.Design.Widget;
 using Android.Views;
 using Cirrious.MvvmCross.Binding.Droid.BindingContext;
-using MvvmCross.Droid.Support.V7.Fragging;
+using MvvmCross.Droid.Support.V7.Fragging.Attributes;
 using MvvmCross.Droid.Support.V7.Fragging.Fragments;
 using Example.Core.ViewModels;
 using Example.Droid.Activities;
