+++ Samples/Example.Droid/packages.config
@@ -1,10 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="MvvmCross.HotTuna.Binding" version="4.0.0-beta5" targetFramework="MonoAndroid50" />
-  <package id="MvvmCross.HotTuna.CrossCore" version="4.0.0-beta5" targetFramework="MonoAndroid50" />
-  <package id="MvvmCross.HotTuna.MvvmCrossLibraries" version="4.0.0-beta5" targetFramework="MonoAndroid50" />
-  <package id="MvvmCross.HotTuna.Plugin.Json" version="4.0.0-beta5" targetFramework="MonoAndroid50" />
-  <package id="MvvmCross.HotTuna.Plugin.Visibility" version="4.0.0-beta5" targetFramework="MonoAndroid50" />
+  <package id="MvvmCross.HotTuna.Binding" version="4.0.0-beta7" targetFramework="MonoAndroid50" />
+  <package id="MvvmCross.HotTuna.CrossCore" version="4.0.0-beta7" targetFramework="MonoAndroid50" />
+  <package id="MvvmCross.HotTuna.MvvmCrossLibraries" version="4.0.0-beta7" targetFramework="MonoAndroid50" />
+  <package id="MvvmCross.HotTuna.Plugin.Json" version="4.0.0-beta7" targetFramework="MonoAndroid50" />
+  <package id="MvvmCross.HotTuna.Plugin.Visibility" version="4.0.0-beta7" targetFramework="MonoAndroid50" />
   <package id="Newtonsoft.Json" version="7.0.1" targetFramework="monoandroid5" />
   <package id="Xamarin.Android.Support.Design" version="23.1.1.0" targetFramework="monoandroid50" />
   <package id="Xamarin.Android.Support.v4" version="23.1.1.0" targetFramework="monoandroid50" />
